@@ -1,22 +1,21 @@
 module github.com/pefish/go-mysql
 
 require (
-	github.com/go-sql-driver/mysql v1.8.1
+	github.com/go-sql-driver/mysql v1.9.3
 	github.com/google/uuid v1.6.0
+	github.com/joho/godotenv v1.5.1
 	github.com/mitchellh/mapstructure v1.5.0
-	github.com/pefish/go-format v0.5.1
-	github.com/pefish/go-interface v0.1.3
-	github.com/pefish/go-test v0.0.4
-	github.com/pefish/go-time v0.3.4
+	github.com/pefish/go-format v0.5.3
+	github.com/pefish/go-interface v0.1.5
+	github.com/pefish/go-test v0.0.5
+	github.com/pefish/go-time v0.3.5
 	github.com/pkg/errors v0.9.1
 )
 
 require (
 	filippo.io/edwards25519 v1.1.0 // indirect
 	github.com/fatih/structs v1.1.0 // indirect
-	golang.org/x/text v0.19.0 // indirect
+	golang.org/x/text v0.30.0 // indirect
 )
 
-go 1.21
-
-toolchain go1.21.0
+go 1.24.0
